pyproject.toml
[project]
name = "vcf-9x-in-box"
version = "1.0.0"
description = "VMware Cloud Foundation 9.x deployment automation for home lab environments"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pyvmomi>=8.0.0",
    "pytz>=2024.1",
    "requests>=2.31.0",
]

[project.scripts]
generate-kickstart = "scripts.generate_kickstart:main"
create-esxi-usb = "scripts.create_esxi_usb:main"
list-vms = "scripts.list_vms:main"
delete-all-vms = "scripts.delete_all_vms:main"
cleanup-vcf-deployment = "scripts.cleanup_vcf_deployment:main"
deploy-vcf-installer = "scripts.deploy_vcf_installer:main"
setup-vcf-installer = "scripts.setup_vcf_installer:main"
validate-vcf-installer = "scripts.validate_vcf_installer_config:main"
fix-vsan-policy = "scripts.fix_vsan_esa_default_storage_policy:main"
fix-vsan-hcl-timestamp = "scripts.fix_vsan_hcl_timestamp:main"
fix-vsan-hcl-bypass = "scripts.fix_vsan_hcl_bypass:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.14.4",
    "types-pyvmomi>=8.0.0.6",
    "types-pyyaml>=6.0.12.20241230",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
